Various fixes, particularly to streaming services, and updates to support latest Twitter API.
* JsonSerialiser now supports long user ids on stream events. Twitter stream event object properties also updated to use longs instead of ints.
* JsonSerialiser now correctly deserialises 'event' stream objects
* JsonSerialiser no longer crashes deserialising stream events that have no target_object value.
* JsonSerialiser now sets new 'RetrieveAt' property on status objects, useful for caching/refresh actions on statuses.
* JsonSerialiser now correctly deserialises cursor collections on WinRT.
* TwitterExtendedEntity now has default entity type of 'Media'.
* ListFollowersOptions now supports Count property.
* ListFriends now supports Count property.
* Fixes to streaming object deserialisation, reduction in number of serialisation objects created during streaming.

+++ src/TweetSharp/TwitterUserStreams.cs
@@ -185,7 +185,7 @@ namespace TweetSharp
 #endif
     public class TwitterUserStreamDeleteStatus : TwitterStreamArtifact
     {
-        public virtual int UserId { get; set; }
+        public virtual long UserId { get; set; }
         public virtual long StatusId { get; set; }
     }
 
@@ -198,16 +198,16 @@ namespace TweetSharp
 #endif
     public class TwitterUserStreamDeleteDirectMessage : TwitterStreamArtifact
     {
-        public virtual int UserId { get; set; }
+        public virtual long UserId { get; set; }
         public virtual long DirectMessageId { get; set; }
     }
-    
-    /// <summary>
-    /// Denotes a stream event, resulting from a user action;
-    /// the source is always the initiating user, the target is always the affected user, and
-    /// the target object depends on the initiating action: statuses, direct messages, and lists.
-    /// </summary>
-    public class TwitterUserStreamEvent : TwitterUserStreamEventBase
+
+		/// <summary>
+		/// Denotes a stream event, resulting from a user action;
+		/// the source is always the initiating user, the target is always the affected user, and
+		/// the target object depends on the initiating action: statuses, direct messages, and lists.
+		/// </summary>
+	public class TwitterUserStreamEvent : TwitterUserStreamEventBase
     {
         public virtual ITwitterModel TargetObject { get; set; }
